# src/main/resources/static/api.yml
openapi: 3.0.0
info:
  title: Holiday Detail API
  description: Endpoints to retrieve, compare, and analyze public holiday data for multiple countries, including recent holidays, non-weekend counts, and common holidays.
  version: 1.0.0
servers:
  - url: http://localhost:8080/
    description: Local development
tags:
  - name: Holiday api
    description: Endpoints for holiday details
paths:
  /api/v1/holidays/recent:
    get:
      tags:
        - Holiday
      operationId: getRecentHolidays
      summary: Last 3 celebrated holidays for a country
      parameters:
        - name: country
          in: query
          required: true
          schema:
            type: string
            pattern: '^[A-Za-z]{2}$'
          example: US
          description: Two letter country code is required
      responses:
        '200':
          description: Recent holidays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Holiday'
        '400':
          description: Invalid country code
  /api/v1/holidays/non-weekend-count:
    get:
      tags:
        - Holiday
      operationId: getNonWeekendHolidayCounts
      summary: Non-weekend holiday counts for countries in a year (sorted descending)
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
          example: 2025
        - name: countries
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            pattern: '^([A-Za-z]{2})(,[A-Za-z]{2})*$'
          example: US,CA,GB
          description: "Comma-separated list of two-letter country codes (ISO 3166-1 alpha-2)."
      responses:
        '200':
          description: Country counts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CountryHolidayCount'
        '400':
          description: Invalid parameters
  /api/v1/holidays/common:
    get:
      tags:
        - Holiday
      operationId: getCommonHolidays
      summary: Deduplicated common holidays between two countries in a year
      parameters:
        - name: year
          in: query
          required: true
          schema:
            type: integer
          example: 2025
        - name: country1
          in: query
          required: true
          schema:
            type: string
            pattern: '^[A-Za-z]{2}$'
          example: US
          description: "First country code. Must not be the same as country2."
        - name: country2
          in: query
          required: true
          schema:
            type: string
            pattern: '^[A-Za-z]{2}$'
          example: CA
          description: "Second country code. Must not be the same as country1."
      responses:
        '200':
          description: Common holidays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CommonHoliday'
        '400':
          description: Invalid parameters
components:
  schemas:
    Holiday:
      type: object
      properties:
        date:
          type: string
          format: date
        name:
          type: string
      required:
        - date
        - name
    CountryHolidayCount:
      type: object
      properties:
        country:
          type: string
        count:
          type: integer
      required:
        - country
        - count
    CommonHoliday:
      type: object
      properties:
        date:
          type: string
          format: date
        localNameCountry1:
          type: string
        localNameCountry2:
          type: string
      required:
        - date
        - localNameCountry1
        - localNameCountry2